requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pypdf
numpy
pandas
google-genai
python-dotenv
aiofiles
